% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{all_inferedspecies}
\alias{all_inferedspecies}
\title{Model Infered Species}
\usage{
all_inferedspecies(sdfs, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{out_dir}{Path to output directory}
}
\value{
Writes csv of infered species to /QC subdir
}
\description{
Function to find the infered species for all replicates of a sesame run SDFobject
}
