% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{condLEVEL_DMLs}
\alias{condLEVEL_DMLs}
\title{Condition-Level DMLs}
\usage{
condLEVEL_DMLs(test_result, CONDITION, LEVEL, pval = 0.05, effSize = 0.05)
}
\arguments{
\item{test_result}{Test results of the DML summary statistics}

\item{CONDITION}{Comparisons to be me made from the colnames of the sample sheet}

\item{LEVEL}{Which level comparison of the condition DML are we looking at}

\item{pval}{Threshold for pvalue cutoff}

\item{effSize}{Threshold for effect size cutoff}
}
\value{
Updated test_result with colum specific for CONDITION and LEVEL that shows Up, Down, and non-sig DMLs
}
\description{
Function to find the DMLs for a specific comparison of condition levels. To be used for plot generation for this specific level of a condition
}
\examples{
condLEVEL_DMLs(test_result, "Group", "Treatment", pval = 0.05, effSize = 0.05)
}
