% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{create_SE}
\alias{create_SE}
\title{Create Summarized Experiment Object}
\usage{
create_SE(betas, sample_sheet_df, formula)
}
\arguments{
\item{betas}{Preprocessed beta values}

\item{sample_sheet_df}{Data frame of read in sample sheet}

\item{formula}{Formula that will be used for DML, but is also used to select columns to color PCA for}
}
\value{
Summarized Experiemnt object to be used for input of DML run
}
\description{
Function to properly create and format summarized experiement for the DML calculation
}
\examples{
create_SE(betas, sample_sheet_df, ~ Group)
}
