% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{detection_rate}
\alias{detection_rate}
\title{Plot Detection Rate}
\usage{
detection_rate(sdfs, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{out_dir}{Path to output directory}
}
\value{
Saves plot to /QC subdir
}
\description{
Function to plot detection rate QC
}
