% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{DML_analysis}
\alias{DML_analysis}
\title{DML Analysis Subassembly}
\usage{
DML_analysis(betas, sample_sheet_df, smry, formula, out_dir)
}
\arguments{
\item{betas}{Beta values of CpGs}

\item{sample_sheet_df}{Data frame of the sample sheet}

\item{smry}{Summary statistics of DML run}

\item{formula}{Formula that will be used for DML}

\item{out_dir}{Path to output directory}
}
\value{
Saving the updated test_result object now containing Up, Down, not-sig DML results for each comparison
}
\description{
Sub-assembly of the SeSAMe street pipeline to analyse and plot basic DML results
}
\examples{
DML_analysis(betas, sample_sheet_df, smry, ~Group, "path/out_dir")
}
