% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{DMR_DMLs}
\alias{DMR_DMLs}
\title{Diff Methylated Region Analysis}
\usage{
DMR_DMLs(se, smry, CONDITION, LEVEL, out_dir)
}
\arguments{
\item{se}{Summarized experiment object output from create_SE}

\item{smry}{Summary statistics from DML run}

\item{CONDITION}{Metavalue realted to colname of sample sheet}

\item{LEVEL}{Which comparison of condition levels to use DML results from}

\item{out_dir}{Path to output directory}
}
\value{
Writes a csv of the DMR results for each CONDITION-LEVEL contrast
}
\description{
Function to run DMR analysis for CONDITION-LEVEL DML CpGs
}
\examples{
DMR_DMLs(se, smry, "Group", "Treatment", "path/out_dir")
}
