% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{extra_SNP_freq}
\alias{extra_SNP_freq}
\title{Extra SNP Freq Plot}
\usage{
extra_SNP_freq(sdfs, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{out_dir}{Path to output directory}
}
\value{
Saves extra SNP freq plots in /QC subdir
}
\description{
Function to plot the extra SNP allele frequencies
}
