% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{get_betas}
\alias{get_betas}
\title{Calculate Beta Values}
\usage{
get_betas(Idat_dir, prep)
}
\arguments{
\item{Idat_dir}{Path to the methylation array data files}

\item{prep}{Character string of the preprocessing code}
}
\value{
Data frame of beta values
}
\description{
Function to read in Idat files and return the calculated beta values
}
\examples{
betas <- get_betas("./Idat_dir", "TQCDPB")
}
