% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{get_sdfs}
\alias{get_sdfs}
\title{Retreive raw intensities}
\usage{
get_sdfs(Idat_dir, prep)
}
\arguments{
\item{Idat_dir}{Path to the methylation array data files}

\item{prep}{Character string of the preprocessing code}
}
\value{
list of data frames for each replicates reporting their intensity values
}
\description{
Function to read in Idats and return intensity values
}
\examples{
sdfs <- get_sdfs("./Idat_dir", "TQCDPB")
}
