% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{GO_analysis_DMLs}
\alias{GO_analysis_DMLs}
\title{GO Analysis of DMLs}
\usage{
GO_analysis_DMLs(test_result, CONDITION, LEVEL, out_dir)
}
\arguments{
\item{test_result}{Test results of DML summary stats}

\item{CONDITION}{Metavalue realted to colname of sample sheet}

\item{LEVEL}{Which comparison of condition levels to use DML results from}

\item{out_dir}{Path to output directory}
}
\value{
Saves a bar plot and Go results csv for up and down GO resepectively
}
\description{
Function to calculate and plot the Gene Ontology results for CONDITION-LEVEL DML CpGs
}
\examples{
GO_analysis_DMLs(test_result, "Group", "Treatment", "path/out_dir")
}
