% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{GO_plot}
\alias{GO_plot}
\title{GO Analysis Plot}
\usage{
GO_plot(gostres, num_to_plot = 5, plot_title = "Top 5 GO Terms per Source")
}
\arguments{
\item{gostres}{GO results from SeSAMe package}

\item{num_to_plot}{Number of top terms per source to plot}

\item{plot_title}{Option to change the plot title}
}
\value{
Ggplot2 formatted bar plot of the top terms enriched from each source
}
\description{
Function to plot GO results
}
\examples{
GO_plot(gostres, 5, "Top 5 GO Terms per Source")
}
