% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{GR_QQ_plots}
\alias{GR_QQ_plots}
\title{Green-Red QQ Plots}
\usage{
GR_QQ_plots(sdfs, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{out_dir}{Path to output directory}
}
\value{
Saves QQ plots to /QC subdir
}
\description{
Function to generate the Green-Red signal QQ plots for each replicate
}
