% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{heatmap_DMLs}
\alias{heatmap_DMLs}
\title{DML Heatmap}
\usage{
heatmap_DMLs(betas, sample_sheet_df, test_result, CONDITION, LEVEL, out_dir)
}
\arguments{
\item{betas}{Preprocessed beta values}

\item{sample_sheet_df}{Data frame of read in sample sheet}

\item{test_result}{Test results of DML summary stats}

\item{CONDITION}{Metavalue realted to colname of sample sheet}

\item{LEVEL}{Which comparison of condition levels to use DML results from}

\item{out_dir}{Path to output directory}
}
\value{
Pheatmap formatted heatmap of DML beta values across replicates colored by selected conditions.
}
\description{
Function to plot heatmap of DMLs frrom specific CONDITION-LEVEL comparison
}
\examples{
heatmap_DMLs(betas, sample_sheet_df, test_result, "Group", "Treatment", "path/out_dir")
}
