% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{intensity_beta_plots}
\alias{intensity_beta_plots}
\title{Beta Intensity Plots}
\usage{
intensity_beta_plots(sdfs, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{out_dir}{Path to output directory}
}
\value{
Saves intensity plots in /QC subdir
}
\description{
Function to plot the beta value intensity for each replicate
}
