% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{MatVar}
\alias{MatVar}
\title{Matrix Variance Calculation}
\usage{
MatVar(x)
}
\arguments{
\item{x}{A matrix where rows are measurements and columns are replicates}
}
\value{
A modified matrix of same input dimensions but z-score values calculated for each row.
}
\description{
Function to calculate the z-scores for each row of a matrix
}
\examples{
betas_zscore <- MatVar(betas)
}
