% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{PCA_variance_plot}
\alias{PCA_variance_plot}
\title{PC Variance Plot}
\usage{
PCA_variance_plot(pca, out_dir)
}
\arguments{
\item{pca}{PCA reults of prcomp from preprocessed beta values}

\item{out_dir}{Path to output directory}
}
\value{
GGplit2 formatted Plot of the explained variance saved to /PCA subdir
}
\description{
Function to plot the PC proportion of explained variance
}
\examples{
PCA_variance_plot(pca, "path/out_dir")
}
