% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{plot_inferedstrains}
\alias{plot_inferedstrains}
\title{Model Infered Strains}
\usage{
plot_inferedstrains(sdf, out_dir)
}
\arguments{
\item{out_dir}{Path to output directory}

\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}
}
\value{
Plots probabilities of infered strains to /QC subdir
}
\description{
Plotting the infered mouse strain for each replicate
}
