% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{plot_PCA}
\alias{plot_PCA}
\title{Plot PC1 & PC2}
\usage{
plot_PCA(pca, condition, sample_sheet_df)
}
\arguments{
\item{pca}{PCA reults of prcomp from preprocessed beta values}

\item{condition}{Character value of a condition name which can be linked to a sample sheet column for coloring of the replicates by the levels of that condition}

\item{sample_sheet_df}{Dataframe of read in sample sheet}
}
\value{
GGplot2 formatted dotplot of PC1 and 2 colored by condition variables
}
\description{
Function to plot the first two PCs and color by some condition value
}
\examples{
plot_PCA(pca, "Group", sample_sheet_df)
}
