% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{qual_metrics}
\alias{qual_metrics}
\title{Quality Metrics}
\usage{
qual_metrics(sdfs, prep, out_dir)
}
\arguments{
\item{sdfs}{list of data frames of intensity values for each replicate; output of get_sdfs}

\item{prep}{Character string of the preprocessing code}

\item{out_dir}{Path to output directory}
}
\value{
Outputs quality metrics to /QC subdir in out_dir
}
\description{
Function to calculate and write quality metics & ranked metrics
}
\examples{
sdfs <- get_sdfs("./Idat_dir", "TQCDPB")
qual_metrics(sdfs, "TQCDPB", "output")
}
