% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{remove_NAs}
\alias{remove_NAs}
\title{Remove NAs}
\usage{
remove_NAs(test_result)
}
\arguments{
\item{test_result}{DF of summaryExtractTest() from DML summary}
}
\value{
a modified data frame with CpGs containing NAs removed
}
\description{
Function to remove CpGs with NAs from test_result
}
\examples{
betas <- remove_NAs(betas)
}
