% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{run_DML}
\alias{run_DML}
\title{Run DML}
\usage{
run_DML(betas, sample_sheet_df, formula, cores, out_dir)
}
\arguments{
\item{betas}{Preprocessed beta values}

\item{sample_sheet_df}{Data frame of read in sample sheet}

\item{formula}{Formula that will be used for DML, but is also used to select columns to color PCA for}

\item{cores}{Number of cores to use for parallelization of DML run}

\item{out_dir}{Path to output directory}
}
\value{
Summary statistics object to be used for downstream analysis
}
\description{
Function to run the DML and summary stats calculation
}
\examples{
run_DML(betas, sample_sheet_df, ~ Group, 8, "path/out_dir")
}
