% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{run_PCA}
\alias{run_PCA}
\title{PCA Subassembly}
\usage{
run_PCA(betas, sample_sheet_df, formula, out_dir)
}
\arguments{
\item{betas}{Preprocessed beta values}

\item{sample_sheet_df}{Data frame of read in sample sheet}

\item{formula}{Formula that will be used for DML, but is also used to select columns to color PCA for}

\item{out_dir}{Path to output directory}
}
\value{
Files and plots saved in the /PCA subdir
}
\description{
A function to run all PCA analysis
}
\examples{
run_PCA(betas, sample_sheet_df, ~ Group, "path/out_dir")
}
