% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{run_QC}
\alias{run_QC}
\title{QC Subassembly}
\usage{
run_QC(Idat_dir, prep, out_dir)
}
\arguments{
\item{Idat_dir}{Path to the methylation array data files}

\item{prep}{Code for preprocessing steps from SeSAMe Street run input argument}

\item{out_dir}{Path to output direcrtory}
}
\value{
Files and plots saved to /QC subdir
}
\description{
Function to run all QC functions
}
\examples{
run_QC("path/Idat_dir", "TQCDPB", "path/out_dir")
}
