% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{SS_samplesheet}
\alias{SS_samplesheet}
\title{Read in Sample Sheet}
\usage{
SS_samplesheet(sample_sheet)
}
\arguments{
\item{sample_sheet}{Path to the sample sheet}
}
\value{
Dataframe of the read in sample sheet data
}
\description{
Function to read in the provided SeSAMe STREET sample sheet in the correct formatting
}
\examples{
sample_sheet_df <- SS_samplesheet("SeSAMe_STREET_Sample_Sheet.xlsx")
}
