% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{test_condition_presence}
\alias{test_condition_presence}
\title{Test precense of a input condition}
\usage{
test_condition_presence(condition, sample_sheet_df)
}
\arguments{
\item{condition}{A character string of one of the variables from formula input}

\item{sample_sheet_df}{Data frame of read-in sample sheet provided for pipeline}
}
\value{
If condition column not found in sample sheet, will stop the pipeline and report error
}
\description{
For each of the conditions supplied by the formula input, this function will check if there is a colname present in the provided sample sheet that matches for further analysis.
}
\examples{
test_condition_presence("Group", sample_sheet)
}
