% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{test_output_architecture}
\alias{test_output_architecture}
\title{Tests Output Architecture}
\usage{
test_output_architecture(out_dir)
}
\arguments{
\item{out_dir}{Path to the output directory}
}
\value{
If architecture incorrect, will stop the pipeline and report error
}
\description{
Tests for output directory architecture to make sure that pipeline can output the necessary documents/plots
}
\examples{
test_out_architecture("./output")
}
