% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeSAMeStr_utils.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot}
\usage{
volcano_plot(test_result, CONDITION, LEVEL, out_dir)
}
\arguments{
\item{test_result}{Test results of the DML summary statistics}

\item{CONDITION}{Comparisons to be me made from the colnames of the sample sheet}

\item{LEVEL}{Which level comparison of the condition DML are we looking at}

\item{out_dir}{Path to output directory}
}
\value{
Ggplot2 formatted volcano plot
}
\description{
Function to plot CpG volcano plot
}
\examples{
volcano_plot(test_result, "Group", "Treatment", "path/out_dir")
}
